Tidy up StringPiece and make it use size_t.

+++ re2/onepass.cc
@@ -339,7 +339,7 @@ done:
     return false;
   for (int i = 0; i < nmatch; i++)
     match[i].set(matchcap[2*i],
-                 static_cast<int>(matchcap[2*i+1] - matchcap[2*i]));
+                 static_cast<size_t>(matchcap[2*i+1] - matchcap[2*i]));
   return true;
 }
